pyproject.toml
[tool.poetry]
name = "moonafly"
version = "2.17.2"
description = "Moonafly isn't just a discord bot."
authors = ["LifeAdventurer <life0adventurer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.12.3"
discord-py = "^2.3.2"
googletrans-py = "^4.0.0"
httpx = "^0.26.0"
psutil = "^5.9.8"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
